requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Caching==2.1.0
Flask-Limiter==3.5.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
PyMySQL==1.1.0

# Security
Werkzeug==3.0.1
python-dotenv==1.0.0
itsdangerous==2.1.2
email-validator==2.1.0
bcrypt==4.1.1

# Forms and Validation
WTForms==3.1.1

# API
flask-restful==0.3.10

# Production Server
gunicorn==21.2.0

# Payment Integration
razorpay==1.4.1

# Caching & Queue
redis==5.0.1
celery==5.3.4

# Utilities
Pillow==10.1.0
python-dateutil==2.8.2
PyJWT==2.8.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
flask-debugtoolbar==0.14.1
